+++ pylf/_core.py
@@ -13,24 +13,23 @@ from pylf import _page
 _INTERNAL_MODE = 'L'  # The mode for internal computation
 _WHITE = 255
 _BLACK = 0
-_AMP = 2  # Amplification for 4X SSAA.
 
 _NEWLINE = '\n'
 
 
-def handwrite(text: str, page_settings: tuple, font, color: str, is_half_char, is_end_char, alpha: tuple,
-              anti_aliasing: bool, worker: int, seed) -> list:
+def handwrite(text: str, page_settings: tuple, font, color: str, is_half_char, is_end_char, alpha: tuple, worker: int,
+              seed) -> list:
     """Do the real stuffs for handwriting simulating."""
-    pages = _draw_text(text, page_settings, font, is_half_char, is_end_char, anti_aliasing, seed)
+    pages = _draw_text(text, page_settings, font, is_half_char, is_end_char, seed)
     if not pages:
         return pages
-    renderer = _Renderer(page_settings, color, alpha, anti_aliasing, seed)
+    renderer = _Renderer(page_settings, color, alpha, seed)
     with multiprocessing.Pool(min(worker, len(pages))) as pool:
         images = pool.map(renderer, pages)
     return images
 
 
-def _draw_text(text: str, page_settings: tuple, font, is_half_char, is_end_char, anti_aliasing: bool, seed) -> list:
+def _draw_text(text: str, page_settings: tuple, font, is_half_char, is_end_char, seed) -> list:
     """Draws the text randomly in black images with white color. Note that (box[3] - box[1]) and (box[2] - box[0]) both
     must be greater than corresponding font_size.
     """
@@ -49,8 +48,11 @@ def _draw_text(text: str, page_settings: tuple, font, is_half_char, is_end_char,
         char = next(chars)
         index = 0
         while True:
-            (size, box, font_size, word_spacing, line_spacing, font_size_sigma, line_spacing_sigma,
-             word_spacing_sigma) = _parse_page_setting(page_settings[index % length], anti_aliasing)
+            page_setting = page_settings[index % length]
+            size, box = page_setting["background"].size, page_setting["box"]
+            font_size, font_size_sigma = page_setting["font_size"], page_setting["font_size_sigma"]
+            word_spacing, word_spacing_sigma = page_setting["word_spacing"], page_setting["word_spacing_sigma"]
+            line_spacing, line_spacing_sigma = page_setting["line_spacing"], page_setting["line_spacing_sigma"]
             left, upper, right, lower = box
             page = _page.Page(_INTERNAL_MODE, size, _BLACK, index)
             draw = page.draw
@@ -81,22 +83,6 @@ def _draw_text(text: str, page_settings: tuple, font, is_half_char, is_end_char,
         return pages
 
 
-def _parse_page_setting(page_setting: dict, anti_aliasing: bool) -> tuple:
-    """A helper function of _draw_text"""
-    size = (tuple(i * _AMP for i in page_setting['background'].size)
-            if anti_aliasing else page_setting['background'].size)
-    box = tuple(i * _AMP for i in page_setting['box']) if anti_aliasing else page_setting['box']
-    font_size = page_setting['font_size'] * _AMP if anti_aliasing else page_setting['font_size']
-    word_spacing = page_setting['word_spacing'] * _AMP if anti_aliasing else page_setting['word_spacing']
-    line_spacing = page_setting['line_spacing'] * _AMP if anti_aliasing else page_setting['line_spacing']
-    font_size_sigma = page_setting['font_size_sigma'] * _AMP if anti_aliasing else page_setting['font_size_sigma']
-    word_spacing_sigma = (page_setting['word_spacing_sigma'] * _AMP
-                          if anti_aliasing else page_setting['word_spacing_sigma'])
-    line_spacing_sigma = (page_setting['line_spacing_sigma'] * _AMP
-                          if anti_aliasing else page_setting['line_spacing_sigma'])
-    return size, box, font_size, word_spacing, line_spacing, font_size_sigma, line_spacing_sigma, word_spacing_sigma
-
-
 def _draw_char(draw, char: str, xy: tuple, font) -> int:
     """Draws a single char with the parameters and white color, and returns the offset."""
     draw.text(xy, char, fill=_WHITE, font=font)
@@ -106,11 +92,10 @@ def _draw_char(draw, char: str, xy: tuple, font) -> int:
 class _Renderer(object):
     """A function-like object rendering the foreground that was drawn text and returning rendered image."""
 
-    def __init__(self, page_settings: tuple, color: str, alpha: tuple, anti_aliasing: bool, seed):
+    def __init__(self, page_settings: tuple, color: str, alpha: tuple, seed):
         self._page_settings = page_settings
         self._color = color
         self._alpha = alpha
-        self._anti_aliasing = anti_aliasing
         self._rand = random.Random()
         if seed is None:
             self._hashed_seed = None
@@ -123,8 +108,6 @@ class _Renderer(object):
         else:
             self._rand.seed(a=self._hashed_seed + page.index)
         self._perturb(page)
-        if self._anti_aliasing:
-            self._downscale(page)
         return self._merge(page)
 
     def _perturb(self, page: _page.Page) -> None:
@@ -169,11 +152,6 @@ class _Renderer(object):
             matrix[i, y] = matrix[i + offset, y]
         for i in range(width - offset, width):
             matrix[i, y] = _BLACK
-
-    @staticmethod
-    def _downscale(page: _page.Page) -> None:
-        """Downscales the image for 4X SSAA."""
-        page.image = page.image.resize(size=(page.width // _AMP, page.height // _AMP), resample=image.BOX)
 
     def _merge(self, page: _page.Page):
         """Merges the foreground and the background and returns merged raw image."""